HSliderを実装した

+++ src/ofxTUIWindow.h
@@ -5,7 +5,7 @@
 #include <ofxTUIBaseWindow.h>
 class ofxTUIWindow : public ofxTUIBaseWindow{
 	public:
-		ofxTUIWindow(const int& h, const int& w, const int& py = 0, const int& px = 0):ofxTUIBaseWindow(h,w,py,px){};
+		ofxTUIWindow(const int& h, const int& w, const float& py = 0, const float& px = 0):ofxTUIBaseWindow(h,w,py,px){};
 		virtual ~ofxTUIWindow(){};
 
 		virtual void update(){};
@@ -13,8 +13,8 @@ class ofxTUIWindow : public ofxTUIBaseWindow{
 		virtual void changedWindowSize(){};
 		virtual void keyPressed(const int& key){};
 		virtual void keyReleased(const int& key){};
-		virtual void mouseMoved(const int& px, const int& py){};
-		virtual void mouseDragged(const int& px, const int& py, const int& button){};
-		virtual void mousePressed(const int& px, const int& py, const int& button){};
-		virtual void mouseReleased(const int& px, const int& py, const int& button){};
+		virtual void mouseMoved(const float& px, const float& py){};
+		virtual void mouseDragged(const float& px, const float& py, const int& button){};
+		virtual void mousePressed(const float& px, const float& py, const int& button){};
+		virtual void mouseReleased(const float& px, const float& py, const int& button){};
 };
